+++ src/Adjustments/Adjusters/SimpleTax.php
@@ -21,11 +21,14 @@ use Vanilo\Adjustments\Models\AdjustmentProxy;
 use Vanilo\Adjustments\Models\AdjustmentTypeProxy;
 use Vanilo\Adjustments\Support\HasWriteableTitleAndDescription;
 use Vanilo\Adjustments\Support\IsLockable;
+use Vanilo\Adjustments\Support\RoundsAdjustments;
+use Vanilo\Support\Rounding;
 
 final class SimpleTax implements Adjuster
 {
     use HasWriteableTitleAndDescription;
     use IsLockable;
+    use RoundsAdjustments;
 
     public function __construct(
         private float $rate,
@@ -61,10 +64,10 @@ final class SimpleTax implements Adjuster
 
     private function calculateTaxAmount(Adjustable $adjustable): float
     {
-        return match ($this->isIncluded) {
+        return $this->roundTax(match ($this->isIncluded) {
             true => $adjustable->preAdjustmentTotal() / (100 +  $this->rate) *  $this->rate,
             false => $adjustable->preAdjustmentTotal() *  $this->rate / 100,
-        };
+        });
     }
 
     private function getModelAttributes(Adjustable $adjustable): array